rename: getAbout.js 위치 변경 및 supabase.rpc 명 수정
